requirements.txt
Babel==2.6.0
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.3
django-phone-login==1.1.3
django-phonenumber-field==1.1.0
django-sendsms==0.2.3
djangorestframework==3.9.0
et-xmlfile==1.0.1
firebase-admin==2.13.0
gcloud==0.17.0
google-api-core==1.5.2
google-auth==1.6.0
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
jdcal==1.4
jws==0.1.3
msgpack==0.5.6
oauth2client==3.0.0
openpyxl==2.5.9
phonenumberslite==8.9.16
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.7.5
python-jwt==2.0.1
pytz==2018.7
requests==2.20.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.1
psycopg2==2.6.2
coreapi==2.3.3